Moved common.py to submodules in ./common
